fix(605) - add metadata handling for kamelet resource
