+++ packages/ui/src/models/visualization/flows/index.ts
@@ -1,4 +1,5 @@
 export * from './camel-catalog.service';
 export * from './pipe-visual-entity';
 export * from './kamelet-binding-visual-entity';
+export * from './kamelet-visual-entity';
 export * from './camel-route-visual-entity';
